[UPDATE] Selección y definición del lenguaje Pascal

+++ src/app/controllers/expresiones/aritmeticas/division.controller.ts
@@ -10,7 +10,7 @@ export class Division extends Nodo {
 
   constructor(opIzquierdo: Nodo, opDerecho: Nodo,
     linea: number, columna: number) {
-    super(Tipo.PRIMITIVO, Tipo.DOUBLE, linea, columna);
+    super(Tipo.PRIMITIVO, Tipo.REAL, linea, columna);
 
     this.opIzquierdo = opIzquierdo;
     this.opDerecho = opDerecho;
@@ -39,9 +39,9 @@ export class Division extends Nodo {
 
       return parseInt(resIzquierdo) / parseInt(resDerecho);
 
-    } else if ((this.opIzquierdo.tipo == Tipo.DOUBLE || this.opIzquierdo.tipo == Tipo.INTEGER)
-      && (this.opDerecho.tipo == Tipo.DOUBLE || this.opDerecho.tipo == Tipo.INTEGER)) {
-      this.tipo = Tipo.DOUBLE;
+    } else if ((this.opIzquierdo.tipo == Tipo.REAL || this.opIzquierdo.tipo == Tipo.INTEGER)
+      && (this.opDerecho.tipo == Tipo.REAL || this.opDerecho.tipo == Tipo.INTEGER)) {
+      this.tipo = Tipo.REAL;
 
       if (resIzquierdo == 0) {
         const excepcion: Excepcion = new Excepcion('Semantico',
