+++ src/app/controllers/expresiones/relacionales/diferencia.controller.ts
@@ -28,8 +28,8 @@ export class Diferencia extends Nodo {
     if (this.opIzquierdo.tipo == Tipo.STRING && this.opDerecho.tipo == Tipo.STRING) {
       return resIzquierdo != resDerecho;
 
-    } else if ((this.opIzquierdo.tipo == Tipo.DOUBLE || this.opIzquierdo.tipo == Tipo.INTEGER)
-      && (this.opDerecho.tipo == Tipo.DOUBLE || this.opDerecho.tipo == Tipo.INTEGER)) {
+    } else if ((this.opIzquierdo.tipo == Tipo.REAL || this.opIzquierdo.tipo == Tipo.INTEGER)
+      && (this.opDerecho.tipo == Tipo.REAL || this.opDerecho.tipo == Tipo.INTEGER)) {
       return parseFloat(resIzquierdo) != parseFloat(resDerecho);
 
     } else {
